requirements_test.txt
pytest==6.1.1
pytest-aiohttp==0.3.0
